// src/Lib/KeyMetrics_V3.d.ts
export default interface KeyMetrics_V3 {
    symbol: string;
    date: string;
    fiscalYear: string;
    period: string;
    reportedCurrency: string;
    marketCap: number;
    enterpriseValue: number;
    evToSales: number;
    evToOperatingCashFlow: number;
    evToFreeCashFlow: number;
    evToEBITDA: number;
    netDebtToEBITDA: number;
    currentRatio: number;
    incomeQuality: number;
    grahamNumber: number;
    grahamNetNet: number;
    taxBurden: number;
    interestBurden: number;
    workingCapital: number;
    investedCapital: number;
    returnOnAssets: number;
    operatingReturnOnAssets: number;
    returnOnTangibleAssets: number;
    returnOnEquity: number;
    returnOnInvestedCapital: number;
    returnOnCapitalEmployed: number;
    earningsYield: number;
    freeCashFlowYield: number;
    capexToOperatingCashFlow: number;
    capexToDepreciation: number;
    capexToRevenue: number;
    salesGeneralAndAdministrativeToRevenue: number;
    researchAndDevelopementToRevenue: number;
    stockBasedCompensationToRevenue: number;
    intangiblesToTotalAssets: number;
    averageReceivables: number;
    averagePayables: number;
    averageInventory: number;
    daysOfSalesOutstanding: number;
    daysOfPayablesOutstanding: number;
    daysOfInventoryOutstanding: number;
    operatingCycle: number;
    cashConversionCycle: number;
    freeCashFlowToEquity: number;
    freeCashFlowToFirm: number;
    tangibleAssetValue: number;
    netCurrentAssetValue: number;
}
//# sourceMappingURL=KeyMetrics_V3.d.ts.map