// src/Lib/HistoricalPriceFull_V3.d.ts
export default interface HistoricalPriceFull_V3 {
    date: string;
    open: number;
    high: number;
    low: number;
    close: number;
    adjClose: number;
    volume: number;
    unadjustedVolume: number;
    change: number;
    changePercent: number;
    vwap: number;
    label: string;
    changeOverTime: number;
}
//# sourceMappingURL=HistoricalPriceFull_V3.d.ts.map